+++ visualization/README_DO_NOT_DELETE.txt
@@ -1,0 +1,3 @@
+This directory previously contained visualization helpers. The main visualizer has been consolidated to the repository root `visualize.py`.
+
+Files here were archived or removed by the maintainer. If you need originals, check the repo history.
